requirements.txt
aes==1.2.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.10.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.2
bleach==6.0.0
bs4==0.0.1
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.2
cryptography==40.0.2
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.97.0
fastjsonschema==2.16.3
filelock==3.12.0
fqdn==1.5.1
frozenlist==1.3.3
gTTS==2.3.1
h11==0.14.0
identify==2.5.24
idna==3.4
imageio==2.26.1
ipykernel==6.21.1
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
lazy_loader==0.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==2.0.5
motor==3.1.2
multidict==6.0.4
mypy==1.3.0
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.0
nodeenv==1.8.0
notebook==6.5.3
notebook_shim==0.2.3
numpy==1.24.2
openai==0.27.2
outcome==1.2.0
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.5.1
pluggy==1.0.0
pre-commit==3.3.2
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyclean==2.7.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.9
pydub==0.25.1
Pygments==2.14.0
pymongo==4.3.3
PyPDF2==3.0.1
pypng==0.20220715.0
pyproject_api==1.5.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
qrcode==7.4.2
qtconsole==5.4.3
QtPy==2.3.1
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-image==0.20.0
scipy==1.10.1
selenium==4.8.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
srt==3.5.2
stack-data==0.6.2
starlette==0.27.0
terminado==0.17.1
tifffile==2023.3.21
tinycss2==1.2.1
tornado==6.3.2
tox==4.5.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
typer==0.9.0
typing_extensions==4.6.2
tzdata==2022.7
tzlocal==4.2
uri-template==1.2.0
urllib3==1.26.15
uvicorn==0.22.0
virtualenv==20.23.0
watchdog==2.3.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.7
wsproto==1.2.0
yarl==1.8.2
